+++ after_deploy.sh
@@ -5,15 +5,14 @@ echo "Announce changes to the world!"
 VERSION=$(tomlq -r '.tool.poetry.version' pyproject.toml)
 echo "Version: $VERSION"
 
-# Send announcement to a standalone test Slack channel.
+# Send announcement to a standalone test webhook.
 docker run \
    --rm \
   -v $PWD:/announcer \
   metaswitch/announcer:$VERSION \
     announce \
-    --slackhook $SLACK_HOOK \
+    --webhook $WEBHOOK \
+    --target teams \
     --changelogversion $VERSION \
     --changelogfile /announcer/CHANGELOG.md \
-    --projectname announcer \
-    --username travis-announcer \
-    --iconemoji party_parrot
+    --projectname announcer